+++ src/App.js
@@ -2,367 +2,17 @@ import React, { useState, useEffect, useCallback, useRef } from 'react';
 import { initializeApp } from 'firebase/app';
 import { getAuth, signInAnonymously, signInWithCustomToken, onAuthStateChanged } from 'firebase/auth';
 import { getFirestore, collection, addDoc, onSnapshot, query, deleteDoc, doc, updateDoc } from 'firebase/firestore';
+import MovieModal from "./components/MovieModal";
+import MenuBar from "./components/MenuBar";
+import MovieManagementPage from "./components/MovieManagementPage";
+import SettingsPage from "./components/SettingsPage";
+import { halls, predefinedMovies } from "./constants";
+import { formatDate, parseDate, getMonday } from "./utils";
 
 // Define Firebase config and app ID for Canvas environment
 const appId = typeof __app_id !== 'undefined' ? __app_id : 'default-app-id';
 const firebaseConfig = typeof __firebase_config !== 'undefined' ? JSON.parse(__firebase_config) : {};
 const initialAuthToken = typeof __initial_auth_token !== 'undefined' ? __initial_auth_token : null;
-
-// Helper function to format date to 'YYYY-MM-DD'
-const formatDate = (date) => date.toISOString().split('T')[0];
-
-// Helper function to parse 'YYYY-MM-DD' to Date object
-const parseDate = (dateString) => {
-    const [year, month, day] = dateString.split('-').map(Number);
-    return new Date(year, month - 1, day);
-};
-
-// Helper function to get the Monday of the week for a given date
-const getMonday = (d) => {
-    d = new Date(d);
-    const day = d.getDay();
-    const diff = d.getDate() - day + (day === 0 ? -6 : 1); // Adjust for Sunday (0 = Sunday, 1 = Monday, etc.)
-    return new Date(d.getFullYear(), d.getMonth(), diff);
-};
-
-// Hardcoded halls for the cinema
-const halls = ['Zaal 1', 'Zaal 2', 'Zaal 3', 'Zaal 4']; // Define your cinema halls here
-
-// Predefined movies (you can expand this list or load from Firestore)
-const predefinedMovies = [
-    { title: 'Kies een film...', genre: '', duration: '' }, // Placeholder
-    { title: 'Dune: Part Two', genre: 'Sci-Fi', duration: 166 },
-    { title: 'Inside Out 2', genre: 'Animatie', duration: 96 },
-    { title: 'Bad Boys: Ride or Die', genre: 'Actie/Komedie', duration: 115 },
-    { title: 'Deadpool & Wolverine', genre: 'Actie/Komedie', duration: 127 },
-    { title: 'Furiosa: A Mad Max Saga', genre: 'Actie/Sci-Fi', duration: 148 },
-    { title: 'Despicable Me 4', genre: 'Animatie/Komedie', duration: 95 },
-    { title: 'The Super Mario Bros. Movie', genre: 'Animatie', duration: 92 },
-    { title: 'Elemental', genre: 'Animatie', duration: 101 },
-    { title: 'The Little Mermaid', genre: 'Fantasie', duration: 135 },
-    { title: 'Joy Ride', genre: 'Komedie', duration: 95 },
-    { title: 'De Oneindig', genre: 'Drama', duration: 110 },
-    { title: 'Ruby Kieuwman', genre: 'Komedie', duration: 98 },
-    { title: 'Spider-Man: Across the Spider-Verse', genre: 'Animatie', duration: 140 },
-    { title: 'Casper en Emma', genre: 'Kinderfilm', duration: 75 },
-    { title: 'Indiana Jones and the Dial of Destiny', genre: 'Actie/Avontuur', duration: 154 },
-    { title: 'Insidious: The Red Door', genre: 'Horror', duration: 107 },
-    { title: 'Juf Roos', genre: 'Kinderfilm', duration: 60 },
-    { title: 'Mission: Impossible - Dead Reckoning Part One', genre: 'Actie', duration: 163 },
-    { title: 'The Flash', genre: 'Actie/Sci-Fi', duration: 144 },
-    { title: 'Oppenheimer', genre: 'Biografie/Drama', duration: 180 },
-    { title: 'Barbie', genre: 'Komedie', duration: 114 },
-    { title: 'Meg 2: The Trench', genre: 'Actie/Sci-Fi', duration: 116 },
-    { title: 'Gran Turismo', genre: 'Actie/Drama', duration: 134 },
-    { title: 'The Nun II', genre: 'Horror', duration: 110 },
-    { title: 'A Haunting in Venice', genre: 'Mysterie/Misdaad', duration: 103 },
-    { title: 'Expend4bles', genre: 'Actie', duration: 103 },
-    { title: 'The Creator', genre: 'Sci-Fi', duration: 133 },
-    { title: 'Saw X', genre: 'Horror', duration: 118 },
-    { title: 'Taylor Swift: The Eras Tour', genre: 'Concertfilm', duration: 165 },
-    { title: 'Five Nights at Freddy\'s', genre: 'Horror', duration: 109 },
-];
-
-// Modal Component for adding/editing movies
-function MovieModal({
-    show,
-    onClose,
-    movieData,
-    onSave,
-    predefinedMovies,
-    halls,
-    message,
-    clearMessage
-}) {
-    const [title, setTitle] = useState('');
-    const [time, setTime] = useState('');
-    const [genre, setGenre] = useState('');
-    const [date, setDate] = useState(formatDate(new Date()));
-    const [duration, setDuration] = useState('');
-    const [selectedHall, setSelectedHall] = useState(halls[0]);
-    const [selectedPredefinedMovie, setSelectedPredefinedMovie] = useState('');
-    const [internalMessage, setInternalMessage] = useState('');
-
-    useEffect(() => {
-        if (show) {
-            // Set initial state for the modal form based on movieData or defaults
-            const currentTitle = movieData?.title || '';
-            setTitle(currentTitle);
-            setTime(movieData?.time || '');
-            setGenre(movieData?.genre || '');
-            setDate(movieData?.date || formatDate(new Date()));
-            setDuration(String(movieData?.duration || ''));
-            setSelectedHall(movieData?.hall || halls[0]);
-            setSelectedPredefinedMovie(currentTitle); // Select in dropdown if it matches
-
-            setInternalMessage(message); // Pass external message to internal
-        }
-    }, [show, movieData, halls, message]);
-
-    // Clear internal message when modal is opened or user interacts
-    useEffect(() => {
-        if (show) {
-            setInternalMessage('');
-            if (clearMessage) clearMessage();
-        }
-    }, [show, clearMessage]);
-
-
-    const handlePredefinedMovieChange = (e) => {
-        const selectedTitle = e.target.value;
-        setSelectedPredefinedMovie(selectedTitle);
-
-        if (selectedTitle === 'Kies een film...') {
-            setTitle('');
-            setGenre('');
-            setDuration('');
-            return;
-        }
-
-        const movie = predefinedMovies.find(m => m.title === selectedTitle);
-        if (movie) {
-            setTitle(movie.title);
-            setGenre(movie.genre);
-            setDuration(String(movie.duration));
-        }
-    };
-
-    const handleSubmit = () => {
-        if (!title || !time || !genre || !date || !duration || !selectedHall) {
-            setInternalMessage('Vul alle velden in (titel, tijd, genre, datum, duur, zaal).');
-            return;
-        }
-        const timeRegex = /^(?:2[0-3]|[01]?[0-9]):[0-5][0-9]$/;
-        if (!timeRegex.test(time)) {
-            setInternalMessage('Tijd moet in HH:MM formaat zijn (bijv. 20:00).');
-            return;
-        }
-        if (isNaN(parseInt(duration, 10)) || parseInt(duration, 10) <= 0) {
-            setInternalMessage('Duur moet een positief getal zijn.');
-            return;
-        }
-
-        onSave({
-            id: movieData ? movieData.id : null,
-            title,
-            time,
-            genre,
-            date,
-            duration: parseInt(duration, 10),
-            hall: selectedHall
-        });
-        // onClose is called by handleSaveMovie in parent
-    };
-
-    if (!show) return null;
-
-    return (
-        <div className="fixed inset-0 bg-black bg-opacity-75 flex items-center justify-center z-50 p-4">
-            <div className="bg-purple-700 bg-opacity-95 p-6 rounded-xl shadow-2xl w-full max-w-md animate-fade-in">
-                <h2 className="text-2xl font-semibold mb-5 text-center">{movieData ? 'Film bewerken' : 'Film toevoegen'}</h2>
-                {internalMessage && (
-                    <div className="bg-yellow-400 text-purple-900 p-3 rounded-md mb-4 text-center shadow-md animate-bounce-in">
-                        {internalMessage}
-                    </div>
-                )}
-                <div className="grid grid-cols-1 gap-4">
-                    <select
-                        value={selectedPredefinedMovie}
-                        onChange={handlePredefinedMovieChange}
-                        className="p-3 rounded-md bg-purple-800 border border-purple-600 text-white focus:outline-none focus:ring-2 focus:ring-purple-400 transition-all duration-200 appearance-none"
-                    >
-                        {predefinedMovies.map(movie => (
-                            <option key={movie.title} value={movie.title}>
-                                {movie.title} {movie.duration ? `(${movie.duration} min)` : ''}
-                            </option>
-                        ))}
-                    </select>
-
-                    <input
-                        type="text"
-                        placeholder="Filmtitel"
-                        value={title}
-                        onChange={(e) => setTitle(e.target.value)}
-                        className="p-3 rounded-md bg-purple-800 border border-purple-600 text-white placeholder-purple-300 focus:outline-none focus:ring-2 focus:ring-purple-400 transition-all duration-200"
-                    />
-                    <input
-                        type="time"
-                        placeholder="Tijd (HH:MM)"
-                        value={time}
-                        onChange={(e) => setTime(e.target.value)}
-                        className="p-3 rounded-md bg-purple-800 border border-purple-600 text-white placeholder-purple-300 focus:outline-none focus:ring-2 focus:ring-purple-400 transition-all duration-200"
-                    />
-                    <input
-                        type="date"
-                        value={date}
-                        onChange={(e) => setDate(e.target.value)}
-                        className="p-3 rounded-md bg-purple-800 border border-purple-600 text-white placeholder-purple-300 focus:outline-none focus:ring-2 focus:ring-purple-400 transition-all duration-200"
-                    />
-                    <input
-                        type="number"
-                        placeholder="Duur (minuten)"
-                        value={duration}
-                        onChange={(e) => setDuration(e.target.value)}
-                        className="p-3 rounded-md bg-purple-800 border border-purple-600 text-white placeholder-purple-300 focus:outline-none focus:ring-2 focus:ring-purple-400 transition-all duration-200"
-                    />
-                    <select
-                        value={selectedHall}
-                        onChange={(e) => setSelectedHall(e.target.value)}
-                        className="p-3 rounded-md bg-purple-800 border border-purple-600 text-white focus:outline-none focus:ring-2 focus:ring-purple-400 transition-all duration-200 appearance-none"
-                    >
-                        {halls.map(hallName => (
-                            <option key={hallName} value={hallName}>{hallName}</option>
-                        ))}
-                    </select>
-
-                    <button
-                        onClick={handleSubmit}
-                        className="w-full bg-indigo-500 hover:bg-indigo-600 active:bg-indigo-700 text-white font-bold py-3 px-6 rounded-lg transition-all duration-300 transform hover:scale-105 shadow-lg focus:outline-none focus:ring-2 focus:ring-indigo-400 mt-4"
-                    >
-                        {movieData ? 'Film bijwerken' : 'Film toevoegen'}
-                    </button>
-                    <button
-                        onClick={onClose}
-                        className="w-full bg-gray-500 hover:bg-gray-600 active:bg-gray-700 text-white font-bold py-3 px-6 rounded-lg transition-all duration-300 transform hover:scale-105 shadow-lg focus:outline-none focus:ring-2 focus:ring-gray-400 mt-2"
-                    >
-                        Annuleren
-                    </button>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-// Menu Bar Component
-function MenuBar({
-    selectedPredefinedMovieForQuickAdd,
-    handlePredefinedMovieForQuickAddChange,
-    handleDragStart,
-    handleOpenMovieModal,
-    predefinedMovies,
-    currentPage,
-    setCurrentPage,
-    viewMode,
-    toggleViewMode
-}) {
-    return (
-        <div className="bg-purple-800 text-white w-full py-3 px-6 shadow-xl flex flex-col sm:flex-row items-center justify-between sticky top-0 z-20">
-            <h1 className="text-xl font-bold mb-2 sm:mb-0 mr-4">Bioscoop Planner</h1>
-            <nav className="flex gap-4 mb-2 sm:mb-0">
-                <button
-                    onClick={() => setCurrentPage('programming')}
-                    className={`px-4 py-2 rounded-lg transition-colors duration-200 ${currentPage === 'programming' ? 'bg-indigo-600' : 'bg-purple-700 hover:bg-purple-600'}`}
-                >
-                    Programmering
-                </button>
-                <button
-                    onClick={() => setCurrentPage('movieManagement')}
-                    className={`px-4 py-2 rounded-lg transition-colors duration-200 ${currentPage === 'movieManagement' ? 'bg-indigo-600' : 'bg-purple-700 hover:bg-purple-600'}`}
-                >
-                    Filmbeheer
-                </button>
-                <button
-                    onClick={() => setCurrentPage('settings')}
-                    className={`px-4 py-2 rounded-lg transition-colors duration-200 ${currentPage === 'settings' ? 'bg-indigo-600' : 'bg-purple-700 hover:bg-purple-600'}`}
-                >
-                    Instellingen
-                </button>
-            </nav>
-            <div className="flex flex-col sm:flex-row items-center gap-3 w-full sm:w-auto mt-2 sm:mt-0">
-                {currentPage === 'programming' && ( // Only show quick-add on Programming page
-                    <>
-                        <select
-                            value={selectedPredefinedMovieForQuickAdd}
-                            onChange={handlePredefinedMovieForQuickAddChange}
-                            className="p-2 rounded-md bg-purple-700 border border-purple-500 text-white focus:outline-none focus:ring-2 focus:ring-purple-400 transition-all duration-200 appearance-none flex-grow"
-                        >
-                            {predefinedMovies.map(movie => (
-                                <option key={movie.title} value={movie.title}>
-                                    {movie.title} {movie.duration ? `(${movie.duration} min)` : ''}
-                                </option>
-                            ))}
-                        </select>
-
-                        {selectedPredefinedMovieForQuickAdd && selectedPredefinedMovieForQuickAdd !== 'Kies een film...' && (
-                            <div
-                                className="draggable-movie-preview px-4 py-2 bg-indigo-600 hover:bg-indigo-700 text-white rounded-md cursor-grab transition-colors duration-200 shadow-md"
-                                draggable="true"
-                                onDragStart={handleDragStart}
-                            >
-                                Sleep film: {selectedPredefinedMovieForQuickAdd}
-                            </div>
-                        )}
-                    </>
-                )}
-                <button
-                    onClick={handleOpenMovieModal}
-                    className="px-4 py-2 bg-indigo-500 hover:bg-indigo-600 active:bg-indigo-700 text-white font-bold rounded-lg transition-all duration-300 transform hover:scale-105 shadow-md flex-shrink-0"
-                >
-                    Film handmatig toevoegen
-                </button>
-                {currentPage === 'programming' && ( // Only show view mode toggle on Programming page
-                    <button
-                        onClick={toggleViewMode}
-                        className="px-4 py-2 bg-green-500 hover:bg-green-600 active:bg-green-700 text-white font-bold rounded-lg transition-all duration-300 transform hover:scale-105 shadow-md flex-shrink-0"
-                    >
-                        Schakel naar {viewMode === 'hall' ? 'Overzicht per dag' : 'Overzicht per zaal'}
-                    </button>
-                )}
-            </div>
-        </div>
-    );
-}
-
-// New component for movie management
-function MovieManagementPage({ movies, handleEditMovie, handleDeleteMovie }) {
-    return (
-        <div className="bg-purple-700 bg-opacity-70 p-6 rounded-xl shadow-2xl w-full max-w-5xl lg:max-w-7xl">
-            <h2 className="text-2xl font-semibold mb-5 text-center">Filmbeheer</h2>
-            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-                {movies.length > 0 ? (
-                    movies.map(movie => (
-                        <div key={movie.id} className="bg-purple-800 p-4 rounded-lg shadow-md flex flex-col justify-between">
-                            <div>
-                                <h3 className="text-lg font-bold">{movie.title}</h3>
-                                <p className="text-purple-300 text-sm">Datum: {movie.date}</p>
-                                <p className="text-purple-300 text-sm">Tijd: {movie.time}</p>
-                                <p className="text-purple-300 text-sm">Zaal: {movie.hall}</p>
-                                <p className="text-purple-300 text-sm">Duur: {movie.duration} min</p>
-                            </div>
-                            <div className="mt-4 flex gap-2">
-                                <button
-                                    onClick={() => handleEditMovie(movie)}
-                                    className="bg-blue-500 hover:bg-blue-600 text-white px-3 py-1 rounded-md text-sm transition-colors"
-                                >
-                                    Bewerk
-                                </button>
-                                <button
-                                    onClick={() => handleDeleteMovie(movie.id)}
-                                    className="bg-red-500 hover:bg-red-600 text-white px-3 py-1 rounded-md text-sm transition-colors"
-                                >
-                                    Verwijder
-                                </button>
-                            </div>
-                        </div>
-                    ))
-                ) : (
-                    <p className="text-center text-purple-300 col-span-full">Geen films gevonden in de programmering.</p>
-                )}
-            </div>
-        </div>
-    );
-}
-
-// New component for settings
-function SettingsPage() {
-    return (
-        <div className="bg-purple-700 bg-opacity-70 p-6 rounded-xl shadow-2xl w-full max-w-5xl lg:max-w-7xl">
-            <h2 className="text-2xl font-semibold mb-5 text-center">Instellingen</h2>
-            <p className="text-lg">Hier komen de instellingen van de applicatie.</p>
-            {/* Future settings like changing start/end hours, default view, etc. */}
-        </div>
-    );
-}
 
 function App() {
     const [movies, setMovies] = useState([]);
