+++ src/constants.js
@@ -1,0 +1,35 @@
+export const halls = ['Zaal 1', 'Zaal 2', 'Zaal 3', 'Zaal 4'];
+
+export const predefinedMovies = [
+    { title: 'Kies een film...', genre: '', duration: '' },
+    { title: 'Dune: Part Two', genre: 'Sci-Fi', duration: 166 },
+    { title: 'Inside Out 2', genre: 'Animatie', duration: 96 },
+    { title: 'Bad Boys: Ride or Die', genre: 'Actie/Komedie', duration: 115 },
+    { title: 'Deadpool & Wolverine', genre: 'Actie/Komedie', duration: 127 },
+    { title: 'Furiosa: A Mad Max Saga', genre: 'Actie/Sci-Fi', duration: 148 },
+    { title: 'Despicable Me 4', genre: 'Animatie/Komedie', duration: 95 },
+    { title: 'The Super Mario Bros. Movie', genre: 'Animatie', duration: 92 },
+    { title: 'Elemental', genre: 'Animatie', duration: 101 },
+    { title: 'The Little Mermaid', genre: 'Fantasie', duration: 135 },
+    { title: 'Joy Ride', genre: 'Komedie', duration: 95 },
+    { title: 'De Oneindig', genre: 'Drama', duration: 110 },
+    { title: 'Ruby Kieuwman', genre: 'Komedie', duration: 98 },
+    { title: 'Spider-Man: Across the Spider-Verse', genre: 'Animatie', duration: 140 },
+    { title: 'Casper en Emma', genre: 'Kinderfilm', duration: 75 },
+    { title: 'Indiana Jones and the Dial of Destiny', genre: 'Actie/Avontuur', duration: 154 },
+    { title: 'Insidious: The Red Door', genre: 'Horror', duration: 107 },
+    { title: 'Juf Roos', genre: 'Kinderfilm', duration: 60 },
+    { title: 'Mission: Impossible - Dead Reckoning Part One', genre: 'Actie', duration: 163 },
+    { title: 'The Flash', genre: 'Actie/Sci-Fi', duration: 144 },
+    { title: 'Oppenheimer', genre: 'Biografie/Drama', duration: 180 },
+    { title: 'Barbie', genre: 'Komedie', duration: 114 },
+    { title: 'Meg 2: The Trench', genre: 'Actie/Sci-Fi', duration: 116 },
+    { title: 'Gran Turismo', genre: 'Actie/Drama', duration: 134 },
+    { title: 'The Nun II', genre: 'Horror', duration: 110 },
+    { title: 'A Haunting in Venice', genre: 'Mysterie/Misdaad', duration: 103 },
+    { title: 'Expend4bles', genre: 'Actie', duration: 103 },
+    { title: 'The Creator', genre: 'Sci-Fi', duration: 133 },
+    { title: 'Saw X', genre: 'Horror', duration: 118 },
+    { title: 'Taylor Swift: The Eras Tour', genre: 'Concertfilm', duration: 165 },
+    { title: "Five Nights at Freddy's", genre: 'Horror', duration: 109 }
+];
