+++ src/index.js
@@ -1,0 +1,2 @@
+import App from './App';
+export default App;
